requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-chess>=1.999
stockfish>=3.28.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
